Use pytest fixture instead of manage.py
* Moves tests outside `securedrop/` dir to sibling dir as described in pytest
  best practices, as a step in the direction of our goal of completely
  separating test and application logic, and to ease implementation of
  https://github.com/freedomofpress/securedrop/pull/1612#issuecomment-285823600.
* Tests are now called by invoking `pytest` directly instead of using
  `securedrop/manage.py`. The pytest-session-level setup and tear down
  functionality now lives in `tests/conftest.py`.
* Fixes #1610. By adding `db_session.remove` to `utils.env.tearDown`, which is
  called by all `TestCase.tearDown` methods in our test suite, session scope is
  tied to test method scope. The next time a `Session` object is needed during
  the execution of a proceeding test method, the `scoped_session` registry
  (`db_session`) will call the `sessionmaker` factory as needed, which will
  create a new `Session` object and a new connection to the freshly re-created
  test database. This new `Session` object will be correctly synced with the
  database state, and test methods will execute with this clean context.
* Calls `Crypto.Random.atfork()` in the `start_journalist_server` and
  `start_source_server` functional test `setUp` functions to re-seed the
  `pycrypto` CSPRNG in the Werkzeug server app child processes. This prevents
  the CSPRNG object from throwing an exception because these child processes
  have the same CSPRNG state as their parent, and will dangerously produce the
  same output.
