--- securedrop/.coveragerc
@@ -1,4 +1,0 @@
-[run]
-branch = True
-source = .
-omit = tests/*
