+++ securedrop/.gitignore
@@ -3,4 +3,3 @@ config.py
 
 .sass-cache
 static/css/*
-
