+++ tests/.coveragerc
@@ -1,0 +1,3 @@
+[run]
+branch = True
+source = ../securedrop
