+++ tests/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from os.path import abspath, dirname, join, realpath
+import sys
+
+sys.path.append(abspath(join(dirname(realpath(__file__)), '..', 'securedrop')))
