+++ tests/conftest.py
@@ -1,0 +1,60 @@
+# -*- coding: utf-8 -*-
+import os
+import shutil
+import signal
+import subprocess
+
+import psutil
+import pytest
+
+os.environ['SECUREDROP_ENV'] = 'test'
+import config
+
+# TODO: the PID file for the redis worker is hard-coded below.
+# Ideally this constant would be provided by a test harness.
+# It has been intentionally omitted from `config.py.example`
+# in order to isolate the test vars from prod vars.
+TEST_WORKER_PIDFILE = '/tmp/securedrop_test_worker.pid'
+
+@pytest.fixture(scope='session')
+def setUptearDown():
+    _start_test_rqworker(config)
+    yield
+    _stop_test_rqworker()
+    _cleanup_test_securedrop_dataroot(config)
+
+
+def _start_test_rqworker(config):
+    if not psutil.pid_exists(_get_pid_from_file(TEST_WORKER_PIDFILE)):
+        tmp_logfile = open('/tmp/test_rqworker.log', 'w')
+        subprocess.Popen(['rqworker', 'test',
+                          '-P', config.SECUREDROP_ROOT,
+                          '--pid', TEST_WORKER_PIDFILE],
+                         stdout=tmp_logfile,
+                         stderr=subprocess.STDOUT)
+
+
+def _stop_test_rqworker():
+    rqworker_pid = _get_pid_from_file(TEST_WORKER_PIDFILE)
+    if rqworker_pid:
+        os.kill(rqworker_pid, signal.SIGTERM)
+        try:
+            os.remove(TEST_WORKER_PIDFILE)
+        except OSError:
+            pass
+
+
+def _get_pid_from_file(pid_file_name):
+    try:
+        return int(open(pid_file_name).read())
+    except IOError:
+        return None
+
+
+def _cleanup_test_securedrop_dataroot(config):
+    # Keyboard interrupts or dropping to pdb after a test failure sometimes
+    # result in the temporary test SecureDrop data root not being deleted.
+    try:
+        shutil.rmtree(config.SECUREDROP_DATA_ROOT)
+    except OSError:
+        pass
