+++ tests/functional/test_submit_and_retrieve_message.py
@@ -26,7 +26,3 @@ class SubmitAndRetrieveMessage(
         self._journalist_logs_in()
         self._journalist_checks_messages()
         self._journalist_downloads_message()
-
-
-if __name__ == "__main__":
-    unittest.main(verbosity=2)
