+++ tests/test_crypto_util.py
@@ -1,12 +1,8 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
-
 import os
 import unittest
 
-# Set environment variable so config.py uses a test environment
 os.environ['SECUREDROP_ENV'] = 'test'
-
 import config
 import crypto_util
 import utils
@@ -27,6 +23,3 @@ class TestCryptoUtil(unittest.TestCase):
             crypto_util.clean('foo bar`') # backtick is not currently allowed
         with self.assertRaises(crypto_util.CryptoException):
             crypto_util.clean('bar baz~') # tilde is not currently allowed
-
-if __name__ == "__main__":
-    unittest.main(verbosity=2)
