+++ tests/utils/env.py
@@ -2,16 +2,18 @@
 """Testing utilities related to setup and teardown of test environment.
 """
 import os
+from os.path import abspath, dirname, exists, isdir, join, realpath
 import shutil
 import subprocess
 
 import gnupg
 
-# Set environment variable so config.py uses a test environment
 os.environ['SECUREDROP_ENV'] = 'test'
 import config
 import crypto_util
-from db import init_db
+from db import init_db, db_session
+
+FILES_DIR = abspath(join(dirname(realpath(__file__)), '..', 'files'))
 
 # TODO: the PID file for the redis worker is hard-coded below.  Ideally this
 # constant would be provided by a test harness.  It has been intentionally
@@ -26,7 +28,7 @@ def create_directories():
     """
     for d in (config.SECUREDROP_DATA_ROOT, config.STORE_DIR,
               config.GPG_KEY_DIR, config.TEMP_DIR):
-        if not os.path.isdir(d):
+        if not isdir(d):
             os.mkdir(d)
 
 
@@ -36,7 +38,8 @@ def init_gpg():
     """
     gpg = gnupg.GPG(homedir=config.GPG_KEY_DIR)
     # Faster to import a pre-generated key than to gen a new one every time.
-    for keyfile in ("test_journalist_key.pub", "test_journalist_key.sec"):
+    for keyfile in (join(FILES_DIR, "test_journalist_key.pub"),
+                    join(FILES_DIR, "test_journalist_key.sec")):
         gpg.import_keys(open(keyfile).read())
     return gpg
 
@@ -49,18 +52,14 @@ def setup():
     # Do tests that should always run on app startup
     crypto_util.do_runtime_tests()
     # Start the Python-RQ worker if it's not already running
-    if not os.path.exists(TEST_WORKER_PIDFILE):
+    if not exists(TEST_WORKER_PIDFILE):
         subprocess.Popen(["rqworker",
                           "-P", config.SECUREDROP_ROOT,
                           "--pid", TEST_WORKER_PIDFILE])
 
 
 def teardown():
-    # As this same statement is run in :meth:`manage._run_in_test_environment`,
-    # there is occasionally a race to remove the SECUREDROP_DATA_ROOT, that
-    # results in test failure during tearDown of fixtures. Keeping this same
-    # statement in both places guarantees cleanup happens in all test-running
-    # scenarios.
+    db_session.remove()
     try:
         shutil.rmtree(config.SECUREDROP_DATA_ROOT)
     except OSError as exc:
